requirements.txt
# Requirements cho Foodpanda Analytics Project
# Người phụ trách: Người 1 - Data Engineer

# Data processing
pandas>=1.5.0
numpy>=1.23.0

# Database
sqlalchemy>=2.0.0
pymysql>=1.0.0
mysql-connector-python>=8.0.0

# Data visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.14.0

# Jupyter Notebook
jupyter>=1.0.0
ipykernel>=6.20.0
notebook>=6.5.0

# Environment variables
python-dotenv>=1.0.0

# Utilities
openpyxl>=3.1.0  # For Excel file support if needed

